lazy-load the discord sdk unless its needed
